pyproject.toml
[project]
name = "mine"
version = "0.1.0"
description = "MINE is a lightweight storage management platform."
authors = [
    {name = "Elson Junio",email = "elsonjuniog@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3 (>=1.42.49,<2.0.0)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "mine-spec @ git+https://github.com/elsonjunio/mine-spec.git@v0.1.2",
]

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
dev = "uvicorn mine_backend.main:app --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "blue (>=0.9.1,<0.10.0)"
]
